Code/AlexaSkill/lambda/requirements.txt
boto3==1.28.78
ask-sdk-core==1.19.0
paho-mqtt